+++ src/LogStatFunctions.jl
@@ -1,8 +1,9 @@
 module LogStatFunctions
 
-using LogExpFunctions: logsumexp, logsubexp
+using LogExpFunctions: logsumexp, logsumexp!, logsubexp
 
 export logmeanexp, logvarexp, logstdexp
+export logmeanexp!, logvarexp!, logstdexp!
 
 """
     logmeanexp(A; dims=:)
@@ -41,6 +42,54 @@ function logstdexp(
         A::AbstractArray; dims = :, corrected::Bool = true, logmean = logmeanexp(A; dims)
     )
     return halve!(logvarexp(A; dims, corrected, logmean))
+end
+
+"""
+    logmeanexp!(out, A)
+
+In-place version of [`logmeanexp`](@ref): computes `log.(mean(exp.(A); dims))` over the
+singleton dimensions of `out`, and writes the result to `out`.
+"""
+function logmeanexp!(out::AbstractArray, A::AbstractArray)
+    logsumexp!(out, A)
+    N = length(A) ÷ length(out)
+    out .-= loglen(out, N)
+    return out
+end
+
+"""
+    logvarexp!(out, A; corrected=true, logmean=logmeanexp!(similar(out), A))
+
+In-place version of [`logvarexp`](@ref): computes `log.(var(exp.(A); dims, corrected))`
+over the singleton dimensions of `out`, and writes the result to `out`.
+"""
+function logvarexp!(
+        out::AbstractArray, A::AbstractArray;
+        corrected::Bool = true, logmean = logmeanexp!(similar(out), A)
+    )
+    logsumexp!(out, 2logsubexp.(A, logmean))
+    N = length(A) ÷ length(out)
+    if corrected
+        out .-= loglen(out, N - 1)
+    else
+        out .-= loglen(out, N)
+    end
+    return out
+end
+
+"""
+    logstdexp!(out, A; corrected=true, logmean=logmeanexp!(similar(out), A))
+
+In-place version of [`logstdexp`](@ref): computes `log.(std(exp.(A); dims, corrected))`
+over the singleton dimensions of `out`, and writes the result to `out`.
+"""
+function logstdexp!(
+        out::AbstractArray, A::AbstractArray;
+        corrected::Bool = true, logmean = logmeanexp!(similar(out), A)
+    )
+    logvarexp!(out, A; corrected, logmean)
+    out ./= 2
+    return out
 end
 
 loglen(R::AbstractArray, N::Integer) = log(convert(real(eltype(R)), N))